+++ ui/eslint.config.js
@@ -84,7 +84,6 @@ const legacyRestrictedImportOffenders = [
   'components/Settings/Registry/Stepper/UrlStepper.tsx',
   'components/SpacesSwitcher/MainDesignsContent.tsx',
   'components/SpacesSwitcher/components.tsx',
-  'components/TroubleshootingModalComponent.tsx',
   'components/UserPreferences/index.tsx',
   'components/ViewSwitch.tsx',
   'components/YamlDialog.tsx',
@@ -211,8 +210,6 @@ const legacyInlineStyleOffenders = [
   'components/AppComponents.tsx',
   'components/BBChart.tsx',
   'components/connections/ConnectionTable.columns.tsx',
-  'components/shared/Modal/ConfirmationModal.tsx',
-  'components/dashboard/UnsavedChangesModal.tsx',
   'components/dashboard/charts/ConnectionCharts.tsx',
   'components/dashboard/charts/DashboardMeshModelGraph.tsx',
   'components/dashboard/charts/KubernetesConnectionChart.tsx',
@@ -236,6 +233,7 @@ const legacyInlineStyleOffenders = [
   'components/dashboard/widgets/getting-started/data.tsx',
   'components/data-formatter/index.tsx',
   'components/DatabaseSummary.tsx',
+  'components/designs/lifecycle/DeployConfirmationModal.tsx',
   'components/designs/lifecycle/DeployStepper.tsx',
   'components/designs/lifecycle/DeploymentSummary.tsx',
   'components/designs/lifecycle/DryRun.tsx',
